Module prisma_query::ast

An abstract syntax tree for SQL queries.

The ast module handles everything related building abstract SQL queries without going into database-level specifics. Everything related to the actual query building is in the visitor module.

For prelude, all important imports are in prisma_query::ast::*.

Structs

Column
A column definition.

Function
A database function definition

JoinData
The JOIN table and conditions.

Ordering
A list of definitions for the ORDER BY statement

Over

Row
A collection of values surrounded by parentheses.

RowNumber

Select
A builder for a SELECT statement.

Table
A table definition

Enums

Compare
For modeling comparison expression

ConditionTree
Tree structures and leaves for condition building.

DatabaseValue
A value we can compare and use in database queries.

Expression
A database expression.

FunctionType
A database function type

Join
A representation of a JOIN statement.

Order
The ordering direction

ParameterizedValue
A value we must parameterize for the prepared statement.

Query
A database query

Traits

Aliasable
An object that can be aliased.

Comparable
An item that can be compared against other values in the database.

Conjuctive
AND, OR and NOT conjuctive implementations.

IntoOrderDefinition
Convert the value into an order definition with order item and direction

Joinable
An item that can be joined.

Orderable
An item that can be used in the ORDER BY statement

Functions

row_number
A number from 1 to n in specified order

Type Definitions

OrderDefinition