Enum prisma_query::ast::ParameterizedValue

pub enum ParameterizedValue<'a> {
    Null,
    Integer(i64),
    Real(f64),
    Text(Cow<'a, str>),
    Boolean(bool),
    Array(Vec<ParameterizedValue<'a>>),
    Json(Value),
    Uuid(Uuid),
    DateTime(DateTime<Utc>),
}

A value we must parameterize for the prepared statement.

Variants
Null

A database null

Integer(i64)

An integer value

Real(f64)

A floating point value

Text(Cow<'a, str>)

A string value

Boolean(bool)

A boolean value

Array(Vec<ParameterizedValue<'a>>)Json(Value)

A JSON value

Uuid(Uuid)DateTime(DateTime<Utc>)
Methods

impl<'a> ParameterizedValue<'a>

pub fn is_null(&self) -> bool

true if the ParameterizedValue is null.

pub fn is_text(&self) -> bool

true if the ParameterizedValue is text.

pub fn as_str(&self) -> Option<&str>

Returns a &str if the value is a text, otherwise None.

pub fn into_string(self) -> Option<String>

Transforms the ParameterizedValue to a String if it's text, otherwise None.

pub fn is_integer(&self) -> bool

true if the ParameterizedValue is an integer.

pub fn as_i64(&self) -> Option<i64>

Returns an i64 if the value is an integer, otherwise None.

pub fn is_real(&self) -> bool

true if the ParameterizedValue is a real value.

pub fn as_f64(&self) -> Option<f64>

Returns a f64 if the value is a real value, otherwise None.

pub fn is_bool(&self) -> bool

true if the ParameterizedValue is a boolean value.

pub fn as_bool(&self) -> Option<bool>

Returns a bool if the value is a boolean, otherwise None.

pub fn is_uuid(&self) -> bool

true if the ParameterizedValue is of UUID type.

pub fn as_uuid(&self) -> Option<Uuid>

Returns an UUID if the value is of UUID type, otherwise None.

pub fn is_datetime(&self) -> bool

true if the ParameterizedValue is a DateTime.

pub fn as_datetime(&self) -> Option<DateTime<Utc>>

Returns a DateTime if the value is a DateTime, otherwise None.

pub fn is_json(&self) -> bool

true if the ParameterizedValue is a JSON value.

pub fn as_json(&self) -> Option<&Value>

Returns a reference to a JSON Value if of Json type, otherwise None.

pub fn into_json(self) -> Option<Value>

Transforms to a JSON Value if of Json type, otherwise None.