Module prisma_query::connector

Structs

Mysql
A connector interface for the MySQL database.

PostgreSql
A connector interface for the PostgreSQL database.

ResultRow
An owned version of a Row in a ResultSet. See ResultRowRef for documentation on data access.

ResultRowRef
A reference to a Row in a ResultSet. The columns can be accessed either through their position or using the column name.

ResultSet
Encapsulates a set of results and their respective column names.

ResultSetIterator
Thin iterator for ResultSet rows. Might become lazy one day.

Sqlite

Transaction

Traits

Queryable
Represents a connection.

ToColumnNames

ToRow

ValueIndex