Trait prisma_query::connector::Queryable

pub trait Queryable {
    fn execute<'a>(&mut self, q: Query<'a>) -> Result<Option<Id>>;

    fn query<'a>(&mut self, q: Query<'a>) -> Result<ResultSet>;

    fn query_raw<'a>(
        &mut self, 
        sql: &str, 
        params: &[ParameterizedValue<'a>]
    ) -> Result<ResultSet>;

    fn execute_raw<'a>(
        &mut self, 
        sql: &str, 
        params: &[ParameterizedValue<'a>]
    ) -> Result<u64>;

    fn turn_off_fk_constraints(&mut self) -> Result<()>;

    fn turn_on_fk_constraints(&mut self) -> Result<()>;

    fn empty_tables<'a>(&mut self, tables: Vec<Table>) -> Result<()> { ... }
}

Represents a connection.

Required methods

fn execute<'a>(&mut self, q: Query<'a>) -> Result<Option<Id>>

Executes the given query and returns the ID of the last inserted row.

This is typically used for mutating queries.

fn query<'a>(&mut self, q: Query<'a>) -> Result<ResultSet>

Executes the given query and returns the result set.

This is typically used for select queries.

fn query_raw<'a>(
&mut self,
sql: &str,
params: &[ParameterizedValue<'a>]
) -> Result<ResultSet>

Executes a query given as SQL, interpolating the given parameters and returning a set of results.

fn execute_raw<'a>(
&mut self,
sql: &str,
params: &[ParameterizedValue<'a>]
) -> Result<u64>

Executes a query given as SQL, interpolating the given parameters and returning the number of affected rows.

fn turn_off_fk_constraints(&mut self) -> Result<()>

Turns off all foreign key constraints.

fn turn_on_fk_constraints(&mut self) -> Result<()>

Turns on all foreign key constraints.

Loading content...
Provided methods

fn empty_tables<'a>(&mut self, tables: Vec<Table>) -> Result<()>

Empties the given set of tables.

Loading content...