Trait prisma_query::connector::Transactional

pub trait Transactional {
    type Error;
    fn with_transaction<F, T>(&self, db: &str, f: F) -> Result<T, Self::Error>
    where
        F: FnOnce(&mut dyn Transaction) -> Result<T, Self::Error>;
}

Associated Types

type Error

Loading content...
Required methods

fn with_transaction<F, T>(&self, db: &str, f: F) -> Result<T, Self::Error> where
F: FnOnce(&mut dyn Transaction) -> Result<T, Self::Error>,

Opens a connection and a transaction, which is valid inside the given handler closure.

The transaction is comitted if the result returned by the handler is Ok. Otherise, the transaction is discarded.

Loading content...