Struct prisma_query::pool::mysql::OptsBuilder

pub struct OptsBuilder { /* fields omitted */ }

Provides a way to build Opts.

// You can create new default builder
let mut builder = OptsBuilder::new();
builder.ip_or_hostname(Some("foo"))
       .db_name(Some("bar"))
       .ssl_opts(Some(("/foo/cert.pem", None::<(String, String)>)));

// Or use existing T: Into<Opts>
let mut builder = OptsBuilder::from_opts(existing_opts);
builder.ip_or_hostname(Some("foo"))
       .db_name(Some("bar"));

Connection URL

Opts also could be constructed using connection URL. See docs on OptsBuilder's methods for the list of options available via URL.

Example:

let connection_url = "mysql://root:password@localhost:3307/mysql?prefer_socket=false";
let pool = my::Pool::new(connection_url).unwrap();

Methods

impl OptsBuilder

pub fn new() -> OptsBuilder

pub fn from_opts<T>(opts: T) -> OptsBuilder where
T: Into<Opts>,

pub fn ip_or_hostname<T>(
&mut self,
ip_or_hostname: Option<T>
) -> &mut OptsBuilder where
T: Into<String>,

Address of mysql server (defaults to 127.0.0.1). Hostnames should also work.

pub fn tcp_port(&mut self, tcp_port: u16) -> &mut OptsBuilder

TCP port of mysql server (defaults to 3306).

pub fn socket<T>(&mut self, socket: Option<T>) -> &mut OptsBuilder where
T: Into<String>,

Socket path on unix or pipe name on windows (defaults to None).

Can be defined using socket connection url parameter.

pub fn user<T>(&mut self, user: Option<T>) -> &mut OptsBuilder where
T: Into<String>,

User (defaults to None).

pub fn pass<T>(&mut self, pass: Option<T>) -> &mut OptsBuilder where
T: Into<String>,

Password (defaults to None).

pub fn db_name<T>(&mut self, db_name: Option<T>) -> &mut OptsBuilder where
T: Into<String>,

Database name (defaults to None).

pub fn read_timeout(
&mut self,
read_timeout: Option<Duration>
) -> &mut OptsBuilder

The timeout for each attempt to read from the server (defaults to None).

Note that named pipe connection will ignore duration's nanos, and also note that it is an error to pass the zero Duration to this method.

pub fn write_timeout(
&mut self,
write_timeout: Option<Duration>
) -> &mut OptsBuilder

The timeout for each attempt to write to the server (defaults to None).

Note that named pipe connection will ignore duration's nanos, and also note that it is likely error to pass the zero Duration to this method.

pub fn tcp_keepalive_time_ms(
&mut self,
tcp_keepalive_time_ms: Option<u32>
) -> &mut OptsBuilder

TCP keep alive time for mysql connection (defaults to None). Available as tcp_keepalive_time_ms url parameter.

Can be defined using tcp_keepalive_time_ms connection url parameter.

pub fn tcp_nodelay(&mut self, nodelay: bool) -> &mut OptsBuilder

Set the TCP_NODELAY option for the mysql connection (defaults to true).

Setting this option to false re-enables Nagle's algorithm, which can cause unusually high latency (~40ms) but may increase maximum throughput. See #132.

pub fn prefer_socket(&mut self, prefer_socket: bool) -> &mut OptsBuilder

Prefer socket connection (defaults to true). Available as prefer_socket url parameter with value true or false.

Will reconnect via socket (on named pipe on windows) after TCP connection to 127.0.0.1 if true.

Will fall back to TCP on error. Use socket option to enforce socket connection.

Can be defined using prefer_socket connection url parameter.

pub fn init<T>(&mut self, init: Vec<T>) -> &mut OptsBuilder where
T: Into<String>,

Commands to execute on each new database connection.

pub fn verify_peer(&mut self, verify_peer: bool) -> &mut OptsBuilder

Only available if ssl feature enabled.

Perform or not ssl peer verification (defaults to false). Available as verify_peer url parameter with value true or false.

Only make sense if ssl_opts is not None.

Can be defined using verify_peer connection url parameter.

pub fn ssl_opts<A, B, C>(&mut self, Option<Option<()>>) -> &mut OptsBuilder

Requires ssl feature

pub fn local_infile_handler(
&mut self,
handler: Option<LocalInfileHandler>
) -> &mut OptsBuilder

Callback to handle requests for local files. These are caused by using LOAD DATA LOCAL INFILE queries. The callback is passed the filename, and a Writeable object to receive the contents of that file. If unset, the default callback will read files relative to the current directory.

pub fn tcp_connect_timeout(
&mut self,
timeout: Option<Duration>
) -> &mut OptsBuilder

Tcp connect timeout (defaults to None). Available as tcp_connect_timeout_ms url parameter.

Can be defined using tcp_connect_timeout_ms connection url parameter.

pub fn bind_address<T>(&mut self, bind_address: Option<T>) -> &mut OptsBuilder where
T: Into<SocketAddr>,

Bind address for a client (defaults to None).

Use carefully. Will probably make pool unusable because of address already in use errors.

pub fn stmt_cache_size<T>(&mut self, cache_size: T) -> &mut OptsBuilder where
T: Into<Option<usize>>,

Number of prepared statements cached on the client side (per connection). Defaults to 10.

Available as stmt_cache_size url parameter.

Call with None to reset to default.

Can be defined using stmt_cache_size connection url parameter.

pub fn compress(&mut self, compress: bool) -> &mut OptsBuilder

If true, then client will ask for compression if server supports it (defaults to false).

Available as compress url parameter with value true or false.

pub fn additional_capabilities(
&mut self,
additional_capabilities: CapabilityFlags
) -> &mut OptsBuilder

Additional client capabilities to set (defaults to empty).

This value will be OR'ed with other client capabilities during connection initialisation.

Note

It is a good way to set something like CLIENT_FOUND_ROWS but you should note that it won't let you to interfere with capabilities managed by other options (like CLIENT_SSL or CLIENT_COMPRESS). Also note that some capabilities are reserved, pointless or may broke the connection, so this option should be used with caution.

pub fn connect_attrs<T1, T2>(
&mut self,
connect_attrs: HashMap<T1, T2, RandomState>
) -> &mut OptsBuilder where
T1: Eq + Into<String> + Hash,
T2: Into<String>,

Connect attributes

This value is sent to the server as custom name-value attributes. You can see them from performance_schema tables: session_account_connect_attrs and session_connect_attrs when all of the following conditions are met.

The server is MySQL 5.6 or later, or MariaDB 10.0 or later.

performance_schema is on.

performance_schema_session_connect_attrs_size is -1 or big enough to store specified attributes.

Note

Attribute names that begin with an underscore (_) are not set by application programs because they are reserved for internal use.

The following attributes are sent in addition to ones set by programs.

name	value

_client_name	The client library name (rust-mysql-simple)

_client_version	The client library version

_os	The operation system (target_os cfg feature)

_pid	The client proces ID

_platform	The machine platform (target_arch cfg feature)

program_name	The first element of std::env::args if program_name isn't set by programs.