Struct prisma_query::pool::postgres::Config

pub struct Config { /* fields omitted */ }

Connection configuration.

Configuration can be parsed from libpq-style connection strings. These strings come in two formats:

Key-Value

This format consists of space-separated key-value pairs. Values which are either the empty string or contain whitespace should be wrapped in '. ' and \ characters should be backslash-escaped.

Keys

user - The username to authenticate with. Required.

password - The password to authenticate with.

dbname - The name of the database to connect to. Defaults to the username.

options - Command line options used to configure the server.

application_name - Sets the application_name parameter on the server.

sslmode - Controls usage of TLS. If set to disable, TLS will not be used. If set to prefer, TLS will be used if available, but not used otherwise. If set to require, TLS will be forced to be used. Defaults to prefer.

host - The host to connect to. On Unix platforms, if the host starts with a / character it is treated as the path to the directory containing Unix domain sockets. Otherwise, it is treated as a hostname. Multiple hosts can be specified, separated by commas. Each host will be tried in turn when connecting. Required if connecting with the connect method.

port - The port to connect to. Multiple ports can be specified, separated by commas. The number of ports must be either 1, in which case it will be used for all hosts, or the same as the number of hosts. Defaults to 5432 if omitted or the empty string.

connect_timeout - The time limit in seconds applied to each socket-level connection attempt. Note that hostnames can resolve to multiple IP addresses, and this limit is applied to each address. Defaults to no timeout.

keepalives - Controls the use of TCP keepalive. A value of 0 disables keepalive and nonzero integers enable it. This option is ignored when connecting with Unix sockets. Defaults to on.

keepalives_idle - The number of seconds of inactivity after which a keepalive message is sent to the server. This option is ignored when connecting with Unix sockets. Defaults to 2 hours.

target_session_attrs - Specifies requirements of the session. If set to read-write, the client will check that the transaction_read_write session parameter is set to on. This can be used to connect to the primary server in a database cluster as opposed to the secondary read-only mirrors. Defaults to all.

Examples

host=localhost user=postgres connect_timeout=10 keepalives=0

host=/var/lib/postgresql,localhost port=1234 user=postgres password='password with spaces'

host=host1,host2,host3 port=1234,,5678 user=postgres target_session_attrs=read-write

Url

This format resembles a URL with a scheme of either postgres:// or postgresql://. All components are optional, and the format accept query parameters for all of the key-value pairs described in the section above. Multiple host/port pairs can be comma-separated. Unix socket paths in the host section of the URL should be percent-encoded, as the path component of the URL specifies the database name.

Examples

postgresql://user@localhost

postgresql://user:password@%2Fvar%2Flib%2Fpostgresql/mydb?connect_timeout=10

postgresql://user@host1:1234,host2,host3:5678?target_session_attrs=read-write

postgresql:///mydb?user=user&host=/var/lib/postgresql

Methods

impl Config

pub fn new() -> Config

Creates a new configuration.

pub fn user(&mut self, user: &str) -> &mut Config

Sets the user to authenticate with.

Required.

pub fn password<T>(&mut self, password: T) -> &mut Config where
T: AsRef<[u8]>,

Sets the password to authenticate with.

pub fn dbname(&mut self, dbname: &str) -> &mut Config

Sets the name of the database to connect to.

Defaults to the user.

pub fn options(&mut self, options: &str) -> &mut Config

Sets command line options used to configure the server.

pub fn application_name(&mut self, application_name: &str) -> &mut Config

Sets the value of the application_name runtime parameter.

pub fn ssl_mode(&mut self, ssl_mode: SslMode) -> &mut Config

Sets the SSL configuration.

Defaults to prefer.

pub fn host(&mut self, host: &str) -> &mut Config

Adds a host to the configuration.

Multiple hosts can be specified by calling this method multiple times, and each will be tried in order. On Unix systems, a host starting with a / is interpreted as a path to a directory containing Unix domain sockets.

pub fn host_path<T>(&mut self, host: T) -> &mut Config where
T: AsRef<Path>,

Adds a Unix socket host to the configuration.

Unlike host, this method allows non-UTF8 paths.

pub fn port(&mut self, port: u16) -> &mut Config

Adds a port to the configuration.

Multiple ports can be specified by calling this method multiple times. There must either be no ports, in which case the default of 5432 is used, a single port, in which it is used for all hosts, or the same number of ports as hosts.

pub fn connect_timeout(&mut self, connect_timeout: Duration) -> &mut Config

Sets the timeout applied to socket-level connection attempts.

Note that hostnames can resolve to multiple IP addresses, and this timeout will apply to each address of each host separately. Defaults to no limit.

pub fn keepalives(&mut self, keepalives: bool) -> &mut Config

Controls the use of TCP keepalive.

This is ignored for Unix domain socket connections. Defaults to true.

pub fn keepalives_idle(&mut self, keepalives_idle: Duration) -> &mut Config

Sets the amount of idle time before a keepalive packet is sent on the connection.

This is ignored for Unix domain sockets, or if the keepalives option is disabled. Defaults to 2 hours.

pub fn target_session_attrs(
&mut self,
target_session_attrs: TargetSessionAttrs
) -> &mut Config

Sets the requirements of the session.

This can be used to connect to the primary server in a clustered database rather than one of the read-only secondary servers. Defaults to Any.

pub fn executor<E>(&mut self, executor: E) -> &mut Config where
E: Executor<Box<dyn Future<Item = (), Error = ()> + 'static + Send>> + 'static + Sync + Send,

Sets the executor used to run the connection futures.

Defaults to a postgres-specific tokio Runtime.

pub fn connect<T>(&self, tls_mode: T) -> Result<Client, Error> where
T: MakeTlsConnect<Socket> + 'static + Send,
<T as MakeTlsConnect<Socket>>::TlsConnect: Send,
<T as MakeTlsConnect<Socket>>::Stream: Send,
<<T as MakeTlsConnect<Socket>>::TlsConnect as TlsConnect<Socket>>::Future: Send,

Opens a connection to a PostgreSQL database.