Struct prisma_query::ResultSet

pub struct ResultSet { /* fields omitted */ }

Encapsulates a set of results and their respective column names.

Methods

impl ResultSet

pub fn new(names: &ColumnNames, values: Vec<ResultRow>) -> ResultSet

Creates a new instance, bound to the given column names and result rows.

pub fn column_index(&self, name: &str) -> Result<usize, Error>

Finds a column index for a name.