Trait prisma_query::Connectional

pub trait Connectional {
    fn with_connection<F, T>(&self, db: &str, f: F) -> QueryResult<T>
    where
        F: FnOnce(&mut dyn Connection) -> QueryResult<T>,
        Self: Sized;

    fn execute_on_connection(
        &self, 
        db: &str, 
        query: Query
    ) -> QueryResult<Option<Id>>;

    fn query_on_connection(
        &self, 
        db: &str, 
        query: Query
    ) -> QueryResult<ResultSet>;

    fn query_on_raw_connection(
        &self, 
        db: &str, 
        sql: &str, 
        params: &[ParameterizedValue]
    ) -> QueryResult<ResultSet>;
}

Required methods

fn with_connection<F, T>(&self, db: &str, f: F) -> QueryResult<T> where
F: FnOnce(&mut dyn Connection) -> QueryResult<T>,
Self: Sized,

Opens a connection, which is valid inside the given handler closure..

This method does not open a transaction, and should used for operations not requiring transactions, e.g. single queries or schema mutations.

fn execute_on_connection(
&self,
db: &str,
query: Query
) -> QueryResult<Option<Id>>

fn query_on_connection(&self, db: &str, query: Query) -> QueryResult<ResultSet>

fn query_on_raw_connection(
&self,
db: &str,
sql: &str,
params: &[ParameterizedValue]
) -> QueryResult<ResultSet>

Loading content...