Trait prisma_query::visitor::Visitor

pub trait Visitor<'a> {
    const C_BACKTICK: &'static str;
    const C_WILDCARD: &'static str;

    fn build<Q>(query: Q) -> (String, Vec<ParameterizedValue<'a>>)
    where
        Q: Into<Query<'a>>;

    fn add_parameter(&mut self, value: ParameterizedValue<'a>);

    fn visit_limit_and_offset(
        &mut self, 
        limit: Option<ParameterizedValue<'a>>, 
        offset: Option<ParameterizedValue<'a>>
    ) -> Option<String>;

    fn visit_insert(&mut self, insert: Insert<'a>) -> String;

    fn parameter_substitution(&self) -> String;

    fn visit_aggregate_to_string(&mut self, value: DatabaseValue<'a>) -> String;

    fn visit_parameterized(&mut self, value: ParameterizedValue<'a>) -> String { ... }

    fn visit_joins(&mut self, joins: Vec<Join<'a>>) -> String { ... }

    fn visit_join_data(&mut self, data: JoinData<'a>) -> String { ... }

    fn visit_select(&mut self, select: Select<'a>) -> String { ... }

    fn visit_update(&mut self, update: Update<'a>) -> String { ... }

    fn visit_delete(&mut self, delete: Delete<'a>) -> String { ... }

    fn delimited_identifiers(parts: Vec<&str>) -> String { ... }

    fn visit_query(&mut self, query: Query<'a>) -> String { ... }

    fn visit_union_all(&mut self, ua: UnionAll<'a>) -> String { ... }

    fn visit_columns(&mut self, columns: Vec<DatabaseValue<'a>>) -> String { ... }

    fn visit_database_value(&mut self, value: DatabaseValue<'a>) -> String { ... }

    fn visit_table(&mut self, table: Table<'a>, include_alias: bool) -> String { ... }

    fn visit_column(&mut self, column: Column<'a>) -> String { ... }

    fn visit_row(&mut self, row: Row<'a>) -> String { ... }

    fn visit_conditions(&mut self, tree: ConditionTree<'a>) -> String { ... }

    fn visit_expression(&mut self, expression: Expression<'a>) -> String { ... }

    fn visit_compare(&mut self, compare: Compare<'a>) -> String { ... }

    fn visit_ordering(&mut self, ordering: Ordering<'a>) -> String { ... }

    fn visit_grouping(&mut self, grouping: Grouping<'a>) -> String { ... }

    fn visit_function(&mut self, fun: Function<'a>) -> String { ... }

    fn visit_partitioning(&mut self, over: Over<'a>) -> String { ... }
}

A function travelling through the query AST, building the final query string and gathering parameters sent to the database together with the query.

Associated Constants

const C_BACKTICK: &'static str

Backtick character to surround identifiers, such as column and table names.

const C_WILDCARD: &'static str

Wildcard character to be used in LIKE queries.

Loading content...
Required methods

fn build<Q>(query: Q) -> (String, Vec<ParameterizedValue<'a>>) where
Q: Into<Query<'a>>,

Convert the given Query to an SQL string and a vector of parameters. When certain parameters are replaced with the C_PARAM character in the query, the vector should contain the parameter value in the right position.

The point of entry for visiting query ASTs.

use prisma_query::{ast::*, visitor::*};

fn main() {
    let query = Select::from_table("cats");
    let (sqlite, _) = Sqlite::build(query.clone());
    let (psql, _) = Postgres::build(query.clone());
    let (mysql, _) = Mysql::build(query.clone());

    assert_eq!("SELECT `cats`.* FROM `cats`", sqlite);
    assert_eq!("SELECT \"cats\".* FROM \"cats\"", psql);
    assert_eq!("SELECT `cats`.* FROM `cats`", mysql);
}

fn add_parameter(&mut self, value: ParameterizedValue<'a>)

When called, the visitor decided to not render the parameter into the query, replacing it with the C_PARAM, calling add_parameter with the replaced value.

fn visit_limit_and_offset(
&mut self,
limit: Option<ParameterizedValue<'a>>,
offset: Option<ParameterizedValue<'a>>
) -> Option<String>

The LIMIT and OFFSET statement in the query

fn visit_insert(&mut self, insert: Insert<'a>) -> String

A walk through an INSERT statement

fn parameter_substitution(&self) -> String

What to use to substitute a parameter in the query.

fn visit_aggregate_to_string(&mut self, value: DatabaseValue<'a>) -> String

What to use to substitute a parameter in the query.

Loading content...
Provided methods

fn visit_parameterized(&mut self, value: ParameterizedValue<'a>) -> String

A visit to a value we parameterize

fn visit_joins(&mut self, joins: Vec<Join<'a>>) -> String

The join statements in the query

fn visit_join_data(&mut self, data: JoinData<'a>) -> String

fn visit_select(&mut self, select: Select<'a>) -> String

A walk through a SELECT statement

fn visit_update(&mut self, update: Update<'a>) -> String

A walk through an UPDATE statement

fn visit_delete(&mut self, delete: Delete<'a>) -> String

A walk through an DELETE statement

fn delimited_identifiers(parts: Vec<&str>) -> String

A helper for delimiting an identifier, surrounding every part with C_BACKTICK and delimiting the values with a .

assert_eq!(
    "`a`.`b`",
    Sqlite::delimited_identifiers(vec!["a".into(), "b".into()])
);

fn visit_query(&mut self, query: Query<'a>) -> String

A walk through a complete Query statement

fn visit_union_all(&mut self, ua: UnionAll<'a>) -> String

A walk through a union of SELECT statements

fn visit_columns(&mut self, columns: Vec<DatabaseValue<'a>>) -> String

The selected columns

fn visit_database_value(&mut self, value: DatabaseValue<'a>) -> String

A visit to a value used in an expression

fn visit_table(&mut self, table: Table<'a>, include_alias: bool) -> String

A database table identifier

fn visit_column(&mut self, column: Column<'a>) -> String

A database column identifier

fn visit_row(&mut self, row: Row<'a>) -> String

A row of data used as an expression

fn visit_conditions(&mut self, tree: ConditionTree<'a>) -> String

A walk through the query conditions

fn visit_expression(&mut self, expression: Expression<'a>) -> String

An expression that can either be a single value, a set of conditions or a comparison call

fn visit_compare(&mut self, compare: Compare<'a>) -> String

A comparison expression

fn visit_ordering(&mut self, ordering: Ordering<'a>) -> String

A visit in the ORDER BY section of the query

fn visit_grouping(&mut self, grouping: Grouping<'a>) -> String

A visit in the GROUP BY section of the query

fn visit_function(&mut self, fun: Function<'a>) -> String

fn visit_partitioning(&mut self, over: Over<'a>) -> String

Loading content...