Crate prisma_query

prisma-query

prisma-query is an AST and database-specific visitors for creating SQL statements.

Under construction and will go through several rounds of changes. Not meant for production use in the current form.

Goals

Query generation when the database and conditions are not known beforehand.

Parameterized queries when possible.

A modular design, separate AST for query building and visitors for different databases.

Database support behind a feature flag.

Non-goals

Database-level type-safety in query building or being an ORM.

Database priorities

SQLite will be the first visitor

PostgreSQL

MySQL

More databases will be decided later.

Examples

Chaining conditions

use prisma_query::{ast::*, visitor::*};

fn main() {
    let conditions = "word"
        .equals("meow")
        .and("age".less_than(10))
        .and("paw".equals("warm"));

    let query = Select::from_table("naukio").so_that(conditions);
    let (sql_str, params) = Sqlite::build(query);

    assert_eq!(
        "SELECT `naukio`.* FROM `naukio` WHERE ((`word` = ? AND `age` < ?) AND `paw` = ?)",
        sql_str,
    );

    assert_eq!(
        vec![
            ParameterizedValue::from("meow"),
            ParameterizedValue::from(10),
            ParameterizedValue::from("warm"),
        ],
        params
    );
}

We can chain the conditions together by calling the corresponding conjuctive with the compares. The parameters returned implement the corresponding output trait from the database adapter for easy passing into a prepared statement, in this case rusqlite.

Building the conditions as a tree

use prisma_query::{ast::*, visitor::*};

fn main() {
    let conditions = ConditionTree::and(
        ConditionTree::or("word".equals("meow"), "age".less_than(10)),
        "paw".equals("warm"),
    );

    let query = Select::from_table("naukio").so_that(conditions);
    let (sql, params) = Sqlite::build(query);

    assert_eq!(
        "SELECT `naukio`.* FROM `naukio` WHERE ((`word` = ? OR `age` < ?) AND `paw` = ?)",
        sql,
    );

    assert_eq!(
        vec![
            ParameterizedValue::from("meow"),
            ParameterizedValue::from(10),
            ParameterizedValue::from("warm"),
        ],
        params
    );
}

In cases where more feasible we want to build a ConditionTree manually from the input, e.g. when mapping data using an Into<ConditionTree> trait.

Modules

ast
An abstract syntax tree for SQL queries.

connector

error

visitor
Visitors for reading an abstract SQL syntax tree, generating the query and gathering parameters in the right order.

Structs

ResultRow

ResultRowWithName
Wraps a result row, so it's columns can be accessed by name.

ResultSet
Encapsulates a set of results and their respective column names.

ResultSetIterator
Thin iterator for ResultSet rows. Might become lazy one day.

Traits

Connectional

ToResultRow

Transactional

Type Definitions

QueryResult