Struct prisma_query::ast::Select

pub struct Select {
    pub table: Option<Box<Table>>,
    pub columns: Vec<DatabaseValue>,
    pub conditions: Option<ConditionTree>,
    pub ordering: Ordering,
    pub limit: Option<usize>,
    pub offset: Option<usize>,
    pub joins: Vec<Join>,
}

A builder for a SELECT statement.

Fields
table: Option<Box<Table>> columns: Vec<DatabaseValue> conditions: Option<ConditionTree> ordering: Ordering limit: Option<usize> offset: Option<usize> joins: Vec<Join>
Methods

impl Select

pub fn from<T>(table: T) -> Self where
T: Into<Table>,

Creates a new SELECT statement from the given table.

let query = Select::from("users");
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` LIMIT -1", sql);

The table can be in multiple parts, defining the database.

let query = Select::from(("crm", "users"));
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `crm`.`users`.* FROM `crm`.`users` LIMIT -1", sql);

It is also possible to use a nested SELECT.

let select = Table::from(Select::default().value(1)).alias("num");
let query = Select::from(select.alias("num"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `num`.* FROM (SELECT ?) AS `num` LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Integer(1)], params);

pub fn value<T>(self, value: T) -> Self where
T: Into<DatabaseValue>,

Selects a static value as the column.

let query = Select::default().value(1);
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT ?", sql);
assert_eq!(vec![ParameterizedValue::Integer(1)], params);

Creating a qualified asterisk to a joined table:

let join = "dogs".on(("dogs", "slave_id").equals(Column::from(("cats", "master_id"))));

let query = Select::from("cats")
    .value(Table::from("cats").asterisk())
    .value(Table::from("dogs").asterisk())
    .inner_join(join);

let (sql, _) = Sqlite::build(query);

assert_eq!(
    "SELECT `cats`.*, `dogs`.* FROM `cats` INNER JOIN `dogs` ON `dogs`.`slave_id` = `cats`.`master_id` LIMIT -1",
    sql
);

pub fn column<T>(self, column: T) -> Self where
T: Into<Column>,

Adds a column to be selected.

let query = Select::from("users")
    .column("name")
    .column(("users", "id"))
    .column(("crm", "users", "foo"));

let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `name`, `users`.`id`, `crm`.`users`.`foo` FROM `users` LIMIT -1", sql);

pub fn columns<T>(self, columns: Vec<T>) -> Self where
T: Into<DatabaseValue>,

A bulk method to select multiple values.

let query = Select::from("users").columns(vec!["foo", "bar"]);
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT ?, ? FROM `users` LIMIT -1", sql);
assert_eq!(vec![
   ParameterizedValue::Text("foo".to_string()),
   ParameterizedValue::Text("bar".to_string())
], params);

pub fn so_that<T>(self, conditions: T) -> Self where
T: Into<ConditionTree>,

Adds WHERE conditions to the query. See Comparable for more examples.

let query = Select::from("users").so_that("foo".equals("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` = ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("bar".to_string())], params);

pub fn inner_join<J>(self, join: J) -> Self where
J: Into<JoinData>,

Adds INNER JOIN clause to the query.

let join = "posts".alias("p").on(("p", "user_id").equals(Column::from(("users", "id"))));
let query = Select::from("users").inner_join(join);
let (sql, _) = Sqlite::build(query);

assert_eq!(
    "SELECT `users`.* FROM `users` INNER JOIN `posts` AS `p` ON `p`.`user_id` = `users`.`id` LIMIT -1",
    sql
);

pub fn left_outer_join<J>(self, join: J) -> Self where
J: Into<JoinData>,

Adds LEFT OUTER JOIN clause to the query.

let join = "posts".alias("p").on(("p", "visible").equals(true));
let query = Select::from("users").left_outer_join(join);
let (sql, params) = Sqlite::build(query);

assert_eq!(
    "SELECT `users`.* FROM `users` LEFT OUTER JOIN `posts` AS `p` ON `p`.`visible` = ? LIMIT -1",
    sql
);

assert_eq!(vec![ParameterizedValue::Boolean(true)], params);

pub fn order_by<T>(self, value: T) -> Self where
T: IntoOrderDefinition,

Adds an ordering to the ORDER BY section.

let query = Select::from("users")
    .order_by("foo")
    .order_by("baz".ascend())
    .order_by("bar".descend());

let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` ORDER BY `foo`, `baz` ASC, `bar` DESC LIMIT -1", sql);

pub fn limit(self, limit: usize) -> Self

Sets the LIMIT value.

let query = Select::from("users").limit(10);
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` LIMIT 10", sql);

pub fn offset(self, offset: usize) -> Self

Sets the OFFSET value.

let query = Select::from("users").offset(10);
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` LIMIT -1 OFFSET 10", sql);