Trait prisma_query::ast::Comparable

pub trait Comparable {
    fn equals<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn not_equals<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn less_than<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn less_than_or_equals<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn greater_than<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn greater_than_or_equals<T>(self, comparison: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn in_selection<T>(self, selection: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn not_in_selection<T>(self, selection: T) -> Compare
    where
        T: Into<DatabaseValue>;

    fn like<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn not_like<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn begins_with<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn not_begins_with<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn ends_into<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn not_ends_into<T>(self, pattern: T) -> Compare
    where
        T: Into<String>;

    fn is_null(self) -> Compare;

    fn is_not_null(self) -> Compare;

    fn between<T, V>(self, left: T, right: V) -> Compare
    where
        T: Into<DatabaseValue>,
        V: Into<DatabaseValue>;

    fn not_between<T, V>(self, left: T, right: V) -> Compare
    where
        T: Into<DatabaseValue>,
        V: Into<DatabaseValue>;
}

An item that can be compared against other values in the database.

Required Methods

fn equals<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if both sides are the same value.

let query = Select::from("users").so_that("foo".equals("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` = ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("bar".to_string())], params);

fn not_equals<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if both sides are not the same value.

let query = Select::from("users").so_that("foo".not_equals("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` <> ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("bar".to_string())], params);

fn less_than<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is smaller than the right side.

let query = Select::from("users").so_that("foo".less_than(10));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` < ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Integer(10)], params);

fn less_than_or_equals<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is smaller than the right side or the same.

let query = Select::from("users").so_that("foo".less_than_or_equals(10));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` <= ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Integer(10)], params);

fn greater_than<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is bigger than the right side.

let query = Select::from("users").so_that("foo".greater_than(10));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` > ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Integer(10)], params);

fn greater_than_or_equals<T>(self, comparison: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is bigger than the right side or the same.

let query = Select::from("users").so_that("foo".greater_than_or_equals(10));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` >= ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Integer(10)], params);

fn in_selection<T>(self, selection: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is included in the right side collection.

let query = Select::from("users").so_that("foo".in_selection(vec![1, 2]));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` IN (?, ?) LIMIT -1", sql);
assert_eq!(vec![
    ParameterizedValue::Integer(1),
    ParameterizedValue::Integer(2),
], params);

fn not_in_selection<T>(self, selection: T) -> Compare where
T: Into<DatabaseValue>,

Tests if the left side is not included in the right side collection.

let query = Select::from("users").so_that("foo".not_in_selection(vec![1, 2]));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` NOT IN (?, ?) LIMIT -1", sql);
assert_eq!(vec![
    ParameterizedValue::Integer(1),
    ParameterizedValue::Integer(2),
], params);

fn like<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side includes the right side string.

let query = Select::from("users").so_that("foo".like("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("%bar%".to_string())], params);

fn not_like<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side does not include the right side string.

let query = Select::from("users").so_that("foo".not_like("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` NOT LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("%bar%".to_string())], params);

fn begins_with<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side starts with the right side string.

let query = Select::from("users").so_that("foo".begins_with("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("bar%".to_string())], params);

fn not_begins_with<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side doesn't start with the right side string.

let query = Select::from("users").so_that("foo".not_begins_with("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` NOT LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("bar%".to_string())], params);

fn ends_into<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side ends into the right side string.

let query = Select::from("users").so_that("foo".ends_into("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("%bar".to_string())], params);

fn not_ends_into<T>(self, pattern: T) -> Compare where
T: Into<String>,

Tests if the left side does not end into the right side string.

let query = Select::from("users").so_that("foo".not_ends_into("bar"));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` NOT LIKE ? LIMIT -1", sql);
assert_eq!(vec![ParameterizedValue::Text("%bar".to_string())], params);

fn is_null(self) -> Compare

Tests if the left side is NULL.

let query = Select::from("users").so_that("foo".is_null());
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` IS NULL LIMIT -1", sql);

fn is_not_null(self) -> Compare

Tests if the left side is not NULL.

let query = Select::from("users").so_that("foo".is_not_null());
let (sql, _) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` IS NOT NULL LIMIT -1", sql);

fn between<T, V>(self, left: T, right: V) -> Compare where
T: Into<DatabaseValue>,
V: Into<DatabaseValue>,

Tests if the value is between two given values.

let query = Select::from("users").so_that("foo".between(420, 666));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` BETWEEN ? AND ? LIMIT -1", sql);

assert_eq!(vec![
    ParameterizedValue::Integer(420),
    ParameterizedValue::Integer(666)
], params);

fn not_between<T, V>(self, left: T, right: V) -> Compare where
T: Into<DatabaseValue>,
V: Into<DatabaseValue>,

Tests if the value is not between two given values.

let query = Select::from("users").so_that("foo".not_between(420, 666));
let (sql, params) = Sqlite::build(query);

assert_eq!("SELECT `users`.* FROM `users` WHERE `foo` NOT BETWEEN ? AND ? LIMIT -1", sql);

assert_eq!(vec![
    ParameterizedValue::Integer(420),
    ParameterizedValue::Integer(666)
], params);