Trait prisma_query::connector::Queryable

pub trait Queryable {
    fn execute(&mut self, q: Query) -> Result<Option<Id>>;

    fn query(&mut self, q: Query) -> Result<ResultSet>;

    fn query_raw(
        &mut self, 
        sql: &str, 
        params: &[ParameterizedValue]
    ) -> Result<ResultSet>;

    fn execute_raw(
        &mut self, 
        sql: &str, 
        params: &[ParameterizedValue]
    ) -> Result<u64>;

    fn turn_off_fk_constraints(&mut self) -> Result<()>;

    fn turn_on_fk_constraints(&mut self) -> Result<()>;

    fn start_transaction<'a>(&'a mut self) -> Result<Transaction<'a>>;

    fn raw_cmd(&mut self, cmd: &str) -> Result<()>;

    fn empty_tables(&mut self, tables: Vec<Table>) -> Result<()> { ... }

    fn insert(&mut self, q: Insert) -> Result<Option<Id>> { ... }

    fn update(&mut self, q: Update) -> Result<()> { ... }

    fn delete(&mut self, q: Delete) -> Result<()> { ... }
}

Represents a connection or a transaction that can be queried.

Required methods

fn execute(&mut self, q: Query) -> Result<Option<Id>>

Executes the given query and returns the ID of the last inserted row.

fn query(&mut self, q: Query) -> Result<ResultSet>

Executes the given query and returns the result set.

fn query_raw(
&mut self,
sql: &str,
params: &[ParameterizedValue]
) -> Result<ResultSet>

Executes a query given as SQL, interpolating the given parameters and returning a set of results.

fn execute_raw(
&mut self,
sql: &str,
params: &[ParameterizedValue]
) -> Result<u64>

Executes a query given as SQL, interpolating the given parameters and returning the number of affected rows.

fn turn_off_fk_constraints(&mut self) -> Result<()>

Turns off all foreign key constraints.

fn turn_on_fk_constraints(&mut self) -> Result<()>

Turns on all foreign key constraints.

fn start_transaction<'a>(&'a mut self) -> Result<Transaction<'a>>

Starts a new transaction

fn raw_cmd(&mut self, cmd: &str) -> Result<()>

Runs a command in the database, for queries that can't be run using prepared statements.

Loading content...
Provided methods

fn empty_tables(&mut self, tables: Vec<Table>) -> Result<()>

Empties the given set of tables.

fn insert(&mut self, q: Insert) -> Result<Option<Id>>

For inserting data. Returns the ID of the last inserted row.

fn update(&mut self, q: Update) -> Result<()>

For updating data.

fn delete(&mut self, q: Delete) -> Result<()>

For deleting data.

Loading content...