Trait prisma_query::visitor::Visitor

pub trait Visitor {
    const C_PARAM: &'static str;
    const C_BACKTICK: &'static str;
    const C_WILDCARD: &'static str;

    fn build<Q>(query: Q) -> (String, Vec<ParameterizedValue>)
    where
        Q: Into<Query>;

    fn add_parameter(&mut self, value: ParameterizedValue);

    fn visit_limit(&mut self, limit: Option<usize>) -> String;

    fn visit_offset(&mut self, offset: usize) -> String;

    fn visit_function(&mut self, fun: Function) -> String;

    fn visit_partitioning(&mut self, fun: Over) -> String;

    fn visit_joins(&mut self, joins: Vec<Join>) -> String { ... }

    fn visit_join_data(&mut self, data: JoinData) -> String { ... }

    fn visit_select(&mut self, select: Select) -> String { ... }

    fn visit_insert(&mut self, insert: Insert) -> String { ... }

    fn visit_update(&mut self, update: Update) -> String { ... }

    fn visit_delete(&mut self, delete: Delete) -> String { ... }

    fn delimited_identifiers(parts: Vec<String>) -> String { ... }

    fn visit_query(&mut self, query: Query) -> String { ... }

    fn visit_columns(&mut self, columns: Vec<DatabaseValue>) -> String { ... }

    fn visit_parameterized(&mut self, value: ParameterizedValue) -> String { ... }

    fn visit_database_value(&mut self, value: DatabaseValue) -> String { ... }

    fn visit_table(&mut self, table: Table, include_alias: bool) -> String { ... }

    fn visit_column(&mut self, column: Column) -> String { ... }

    fn visit_row(&mut self, row: Row) -> String { ... }

    fn visit_conditions(&mut self, tree: ConditionTree) -> String { ... }

    fn visit_expression(&mut self, expression: Expression) -> String { ... }

    fn visit_compare(&mut self, compare: Compare) -> String { ... }

    fn visit_ordering(&mut self, ordering: Ordering) -> String { ... }
}

A function travelling through the query AST, building the final query string and gathering parameters sent to the database together with the query.

Associated Constants

const C_PARAM: &'static str

Parameter character when parameterizing values in the query.

const C_BACKTICK: &'static str

Backtick character to surround identifiers, such as column and table names.

const C_WILDCARD: &'static str

Wildcard character to be used in LIKE queries.

Required Methods

fn build<Q>(query: Q) -> (String, Vec<ParameterizedValue>) where
Q: Into<Query>,

Convert the given Query to an SQL string and a vector of parameters. When certain parameters are replaced with the C_PARAM character in the query, the vector should contain the parameter value in the right position.

fn add_parameter(&mut self, value: ParameterizedValue)

When called, the visitor decided to not render the parameter into the query, replacing it with the C_PARAM, calling add_parameter with the replaced value.

fn visit_limit(&mut self, limit: Option<usize>) -> String

The LIMIT statement in the query

fn visit_offset(&mut self, offset: usize) -> String

The OFFSET statement in the query

fn visit_function(&mut self, fun: Function) -> String

A database function.

fn visit_partitioning(&mut self, fun: Over) -> String

A partitioning statement.

Provided Methods

fn visit_joins(&mut self, joins: Vec<Join>) -> String

The join statements in the query

fn visit_join_data(&mut self, data: JoinData) -> String

fn visit_select(&mut self, select: Select) -> String

A walk through a SELECT statement

fn visit_insert(&mut self, insert: Insert) -> String

A walk through an INSERT statement

fn visit_update(&mut self, update: Update) -> String

A walk through an UPDATE statement

fn visit_delete(&mut self, delete: Delete) -> String

A walk through an DELETE statement

fn delimited_identifiers(parts: Vec<String>) -> String

A helper for delimiting an identifier, surrounding every part with C_BACKTICK and delimiting the values with a .

assert_eq!(
    "`a`.`b`",
    Sqlite::delimited_identifiers(vec!["a".to_string(), "b".to_string()])
);

fn visit_query(&mut self, query: Query) -> String

A walk through a complete Query statement

fn visit_columns(&mut self, columns: Vec<DatabaseValue>) -> String

The selected columns

fn visit_parameterized(&mut self, value: ParameterizedValue) -> String

A visit to a value we parameterize and replace with a ?

fn visit_database_value(&mut self, value: DatabaseValue) -> String

A visit to a value used in an expression

fn visit_table(&mut self, table: Table, include_alias: bool) -> String

A database table identifier

fn visit_column(&mut self, column: Column) -> String

A database column identifier

fn visit_row(&mut self, row: Row) -> String

A row of data used as an expression

fn visit_conditions(&mut self, tree: ConditionTree) -> String

A walk through the query conditions

fn visit_expression(&mut self, expression: Expression) -> String

An expression that can either be a single value, a set of conditions or a comparison call

fn visit_compare(&mut self, compare: Compare) -> String

A comparison expression

fn visit_ordering(&mut self, ordering: Ordering) -> String

A visit in the ORDER BY section of the query